[patch] gapless grid layout
This patch is an altered gridmode layout for dwm, which arranges the windows in a grid.
Instead of using a regular grid, which might leave empty cells when there are not enough
windows to fill the grid, it adjusts the number of windows in the first few columns
to avoid empty cells.

Source: https://dwm.suckless.org/patches/gaplessgrid/